- Updated rw::common::Log and created new log output macros. - Removed a bunch of warning messages
git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/trunk@3624 494c09c7-a67d-49d4-9938-151fb1965be8

+++ RobWork/src/rw/common/IOUtil.cpp
@@ -281,7 +281,7 @@ std::vector<std::string> IOUtil::getFilesInFolder(const std::string& path, bool 
 }
 
 
-int IOUtil::getFileSize(const std::string& filename){
+size_t IOUtil::getFileSize(const std::string& filename){
     namespace fs = boost::filesystem;
     fs::path p( filename, fs::native );
     return fs::file_size( p );
